SNOW-1883649 Added wiremock runner
