+++ test/integration/authentication_mock/testOauthPat.js
@@ -4,55 +4,9 @@ const net = require('net');
 const connParameters = require('../../authentication/connectionParameters');
 const axios = require('axios');
 const AuthTest = require('../../authentication/authTestsBaseClass');
-const WireMockRestClient =  require('wiremock-rest-client').WireMockRestClient;
-const { exec } = require('child_process');
-const Os = require("os");
+const { runWireMockAsync } = require('./wiremockRunner');
 
-
-
-async function runWireMockAsync(port) {
-  let timeoutHandle;
-  const waitingWireMockPromise =  new Promise(async (resolve, reject) => {
-    try {
-      exec(`npx wiremock --enable-browser-proxying --proxy-pass-through  false --port ${port} `);
-      const wireMock = new WireMockRestClient(`http://localhost:${port}`);
-      const readyWireMock = await waitForWiremockStarted(wireMock);
-      resolve(readyWireMock);
-    } catch (err) {
-      reject(err);
-    }
-  });
-
-  const timeout = new Promise((resolve, reject) =>
-    timeoutHandle = setTimeout(
-      () => reject('Wiremock unavailable  after 26000 ms.'),
-      60000));
-  return Promise.race([waitingWireMockPromise, timeout])
-    .then(result => {
-      clearTimeout(timeoutHandle);
-      return result;
-    });
-}
-
-async function waitForWiremockStarted(wireMock) {
-  return fetch(wireMock.baseUri)
-    .then(async (resp) => {
-      if (resp.ok) {
-        return Promise.resolve(wireMock);
-      } else {
-        await new Promise(resolve => setTimeout(resolve, 1000));
-        console.log(`Retry connection to WireMock after wrong response status: ${resp.status}`);
-        return await waitForWiremockStarted(wireMock);
-      }
-    })
-    .catch(async (err) => {
-      await new Promise(resolve => setTimeout(resolve, 1000));
-      console.log(`Retry connection to WireMock after error: ${err}`);
-      return await waitForWiremockStarted(wireMock);
-    });
-}
-
-if (Os.platform() === 'linux') {
+if (true) {
   describe('Wiremock test', function () {
     let port, wireMock;
     before(async () => {
@@ -98,7 +52,7 @@ if (Os.platform() === 'linux') {
 
     it('Successful flow scenario PAT as token', async function () {
       await addWireMockMappingsFromFile('wiremock/mappings/pat/successful_flow.json');
-      const connectionOption = {...connParameters.oauthPATOnWiremock, token: 'MOCK_TOKEN', port: port};
+      const connectionOption = { ...connParameters.oauthPATOnWiremock, token: 'MOCK_TOKEN', port: port };
       authTest.createConnection(connectionOption);
       await authTest.connectAsync();
       authTest.verifyNoErrorWasThrown();
@@ -106,7 +60,7 @@ if (Os.platform() === 'linux') {
 
     it('Successful flow scenario PAT as password', async function () {
       await addWireMockMappingsFromFile('wiremock/mappings/pat/successful_flow.json');
-      const connectionOption = {...connParameters.oauthPATOnWiremock, password: 'MOCK_TOKEN', port: port};
+      const connectionOption = { ...connParameters.oauthPATOnWiremock, password: 'MOCK_TOKEN', port: port };
       authTest.createConnection(connectionOption);
       await authTest.connectAsync();
       authTest.verifyNoErrorWasThrown();
@@ -114,7 +68,7 @@ if (Os.platform() === 'linux') {
 
     it('Invalid token', async function () {
       await addWireMockMappingsFromFile('wiremock/mappings/pat/invalid_pat_token.json');
-      const connectionOption = {...connParameters.oauthPATOnWiremock, token: 'INVALID_TOKEN', port: port};
+      const connectionOption = { ...connParameters.oauthPATOnWiremock, token: 'INVALID_TOKEN', port: port };
       authTest.createConnection(connectionOption);
       await authTest.connectAsync();
       authTest.verifyErrorWasThrown('Programmatic access token is invalid.');
